Upgrade to my linting 1.2.5 rule set

+++ example/lib/example5/main.dart
@@ -905,7 +905,7 @@ class _HomePageState extends State<HomePage> {
                         label: widget.whiteBlend.toString(),
                         value: widget.whiteBlend.toDouble(),
                         onChanged: (double value) {
-                          widget.onWhiteBlendChanged(value.floor().toInt());
+                          widget.onWhiteBlendChanged(value.floor());
                         },
                       ),
                       trailing: Padding(
